# storage/api-docs/api-docs.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Assessment Question Two API",
    "version": "1.0.0",
    "description": "API documentation for authentication and product management."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000/api",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "sanctum": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Token",
        "description": "Enter token in format (Bearer <token>)"
      }
    },
    "schemas": {
      "AuthEnvelope": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "code": { "type": "integer", "example": 200 },
          "message": { "type": "string", "example": "Login successful." },
          "data": {
            "type": "object",
            "properties": {
              "user": {
                "type": "object",
                "properties": {
                  "id": { "type": "integer", "example": 1 },
                  "name": { "type": "string", "example": "John Doe" },
                  "email": { "type": "string", "example": "john@example.com" }
                }
              },
              "token": { "type": "string", "example": "1|your-token-here" },
              "token_type": { "type": "string", "example": "Bearer" }
            }
          }
        }
      },
      "UserEnvelope": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "code": { "type": "integer", "example": 200 },
          "data": {
            "type": "object",
            "properties": {
              "id": { "type": "integer", "example": 1 },
              "name": { "type": "string", "example": "John Doe" },
              "email": { "type": "string", "example": "john@example.com" },
              "email_verified_at": { "type": "string", "format": "date-time", "nullable": true },
              "created_at": { "type": "string", "format": "date-time" },
              "updated_at": { "type": "string", "format": "date-time" }
            }
          }
        }
      },
      "ProductResource": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "name": { "type": "string", "example": "Gaming Mouse" },
          "slug": { "type": "string", "example": "gaming-mouse" },
          "sku": { "type": "string", "example": "MOU-001" },
          "description": { "type": "string", "example": "Wireless gaming mouse", "nullable": true },
          "price": { "type": "string", "example": "99.90" },
          "sale_price": { "type": "string", "example": "79.90", "nullable": true },
          "stock_quantity": { "type": "integer", "example": 20 },
          "image_path": { "type": "string", "nullable": true },
          "is_active": { "type": "boolean", "example": true },
          "category": {
            "type": "object",
            "properties": {
              "id": { "type": "integer", "example": 1 },
              "name": { "type": "string", "example": "Electronics" },
              "slug": { "type": "string", "example": "electronics" }
            }
          },
          "suppliers": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": { "type": "integer", "example": 1 },
                "name": { "type": "string", "example": "Alpha Supply" },
                "email": { "type": "string", "example": "alpha@example.com" }
              }
            }
          },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
      "ProductEnvelope": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "code": { "type": "integer", "example": 200 },
          "message": { "type": "string", "example": "Product created successfully." },
          "data": { "$ref": "#/components/schemas/ProductResource" }
        }
      },
      "ProductListEnvelope": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "code": { "type": "integer", "example": 200 },
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ProductResource" }
          }
        }
      }
    }
  },
  "paths": {
    "/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register a user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "password", "password_confirmation"],
                "properties": {
                  "name": { "type": "string", "example": "John Doe" },
                  "email": { "type": "string", "example": "john@example.com" },
                  "password": { "type": "string", "example": "YOUR_PASSWORD" },
                  "password_confirmation": { "type": "string", "example": "YOUR_PASSWORD" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthEnvelope" }
              }
            }
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login and get token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "example": "john@example.com" },
                  "password": { "type": "string", "example": "YOUR_PASSWORD" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthEnvelope" }
              }
            }
          }
        }
      }
    },
    "/me": {
      "get": {
        "tags": ["Auth"],
        "summary": "Get current authenticated user",
        "security": [{ "sanctum": [] }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserEnvelope" }
              }
            }
          }
        }
      }
    },
    "/logout": {
      "post": {
        "tags": ["Auth"],
        "summary": "Logout current user",
        "security": [{ "sanctum": [] }],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/products": {
      "get": {
        "tags": ["Products"],
        "summary": "List products",
        "security": [{ "sanctum": [] }],
        "parameters": [
          { "name": "category_id", "in": "query", "schema": { "type": "integer" } },
          { "name": "min_price", "in": "query", "schema": { "type": "number", "format": "float" } },
          { "name": "max_price", "in": "query", "schema": { "type": "number", "format": "float" } },
          { "name": "stock_level", "in": "query", "schema": { "type": "string", "enum": ["in_stock", "out_of_stock", "low_stock"] } },
          { "name": "low_stock_threshold", "in": "query", "schema": { "type": "integer" } },
          { "name": "per_page", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProductListEnvelope" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Products"],
        "summary": "Create product",
        "security": [{ "sanctum": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["category_id", "name", "slug", "sku", "price", "stock_quantity"],
                "properties": {
                  "category_id": { "type": "integer", "example": 1 },
                  "supplier_ids": { "type": "array", "items": { "type": "integer" }, "example": [1, 2] },
                  "name": { "type": "string", "example": "Gaming Mouse" },
                  "slug": { "type": "string", "example": "gaming-mouse" },
                  "sku": { "type": "string", "example": "MOU-001" },
                  "description": { "type": "string", "example": "Wireless gaming mouse" },
                  "price": { "type": "number", "format": "float", "example": 99.9 },
                  "sale_price": { "type": "number", "format": "float", "nullable": true, "example": 79.9 },
                  "stock_quantity": { "type": "integer", "example": 20 },
                  "image_path": { "type": "string", "nullable": true },
                  "is_active": { "type": "boolean", "example": true }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProductEnvelope" }
              }
            }
          }
        }
      }
    },
    "/products/{product}": {
      "get": {
        "tags": ["Products"],
        "summary": "Show product",
        "security": [{ "sanctum": [] }],
        "parameters": [
          { "name": "product", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProductEnvelope" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Products"],
        "summary": "Update product",
        "security": [{ "sanctum": [] }],
        "parameters": [
          { "name": "product", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "category_id": { "type": "integer", "example": 2 },
                  "supplier_ids": { "type": "array", "items": { "type": "integer" }, "example": [2] },
                  "name": { "type": "string", "example": "Gaming Mouse Pro" },
                  "slug": { "type": "string", "example": "gaming-mouse-pro" },
                  "sku": { "type": "string", "example": "MOU-001" },
                  "description": { "type": "string", "example": "Updated mouse" },
                  "price": { "type": "number", "format": "float", "example": 109.9 },
                  "sale_price": { "type": "number", "format": "float", "nullable": true },
                  "stock_quantity": { "type": "integer", "example": 15 },
                  "image_path": { "type": "string", "nullable": true },
                  "is_active": { "type": "boolean", "example": true }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProductEnvelope" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Products"],
        "summary": "Delete product",
        "security": [{ "sanctum": [] }],
        "parameters": [
          { "name": "product", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "code": { "type": "integer", "example": 200 },
                    "message": { "type": "string", "example": "Product deleted successfully." }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}